some fixes, mostly profiling

+++ source/poTween.h
@@ -21,8 +21,8 @@ public:
 	typedef boost::function<void()> poTweenFinishedCallback;
 	
 	typedef bool (*poTweenFunction)(float *val, float begin_val, float end_val,
-									float time, float begin_time, float end_time, float duration,
-									float extra1, float extra2);
+									double time, double begin_time, double end_time, double duration,
+									double extra1, double extra2);
 	
 	poTweenBase();
 	virtual ~poTweenBase();
@@ -33,9 +33,9 @@ public:
 	poTweenBase& setNotification(poObject *obj, const std::string &msg, const poDictionary &dict=poDictionary());
 	poTweenBase& clearNotifications();
 	poTweenBase& setTweenFunction(poTweenFunction func);
-	poTweenBase& setDuration(float duration);
-	poTweenBase& setDelay(float delay);
-	poTweenBase& setExtraValues(float e1, float e2);
+	poTweenBase& setDuration(double duration);
+	poTweenBase& setDelay(double delay);
+	poTweenBase& setExtraValues(double e1, double e2);
 	
 	poTweenBase& start();
 	poTweenBase& stop(bool jump_to_end=false);
@@ -54,8 +54,8 @@ protected:
 	virtual void swapBeginAndEnd() = 0;
 	virtual void slewBeginValue() = 0;
 	// returns true when the tween is complete
-	virtual bool updateTweenWithTime(float time, float begin, float end, 
-									 float duration, float e1, float e2) = 0;
+	virtual bool updateTweenWithTime(double time, double begin, double end, 
+									 double duration, double e1, double e2) = 0;
 	
 	poTweenFunction tween_func;
 	
@@ -71,9 +71,9 @@ private:
 	int repeat_count, repeat_counter;
 	poTweenRepeat repeat_type;
 	
-	float time, last_time;
-	float begin_time, end_time, duration, delay;
-	float extra1, extra2;
+	double time, last_time;
+	double begin_time, end_time, duration, delay;
+	double extra1, extra2;
 	
 	std::vector<poTweenFinishedCallback> callbacks;
 };
@@ -81,16 +81,16 @@ private:
 
 static bool tweenUpdater(poTweenBase::poTweenFunction func, 
 						 float *value, float begin_value, float end_value,
-						 float time, float begin, float end,
-						 float duration, float e1, float e2) 
+						 double time, double begin, double end,
+						 double duration, double e1, double e2) 
 {
 	return func(value, begin_value, end_value, time, begin, end, duration, e1, e2);
 }
 
 static bool tweenUpdater(poTweenBase::poTweenFunction func, 
 						 poPoint *value, poPoint begin_value, poPoint end_value,
-						 float time, float begin, float end,
-						 float duration, float e1, float e2) 
+						 double time, double begin, double end,
+						 double duration, double e1, double e2) 
 {
     bool A = func(&value->x, begin_value.x, end_value.x, time, begin, end, duration, e1, e2);
     bool B = func(&value->y, begin_value.y, end_value.y, time, begin, end, duration, e1, e2);
@@ -100,8 +100,8 @@ static bool tweenUpdater(poTweenBase::poTweenFunction func,
 
 static bool tweenUpdater(poTweenBase::poTweenFunction func, 
 						 poColor *value, poColor begin_value, poColor end_value,
-						 float time, float begin, float end,
-						 float duration, float e1, float e2) 
+						 double time, double begin, double end,
+						 double duration, double e1, double e2) 
 {
 	poHSVColor hsv1, hsv2(begin_value), hsv3(end_value);
 	bool d1 = func(&hsv1.H, hsv2.H, hsv3.H, time, begin, end, duration, e1, e2);
@@ -131,8 +131,8 @@ protected:
 	void swapBeginAndEnd()	{std::swap(begin, end);}
 	void slewBeginValue()	{begin = *value;}
 
-	virtual bool updateTweenWithTime(float t, float b, float e, 
-									 float d, float e1, float e2) 
+	virtual bool updateTweenWithTime(double t, double b, double e, 
+									 double d, double e1, double e2) 
 	{
 		return tweenUpdater(tween_func, value, begin, end, t, b, e, d, e1, e2);
 	}
